legacy/requirements.txt
# Core dependencies
biopython==1.81
pandas==2.0.3
numpy==1.24.3

# Visualization
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2

# Machine Learning
scikit-learn==1.3.0
umap-learn==0.5.4

# Deep Learning
torch==2.1.0
torchvision==0.16.0

# Web Framework
streamlit==1.28.0

# Additional utilities
scipy==1.11.2
requests==2.31.0
